backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.5

# AI / LLM
openai==1.12.0
groq>=0.4.0
google-generativeai==0.8.3
sentence-transformers>=2.2.0
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.21
tiktoken==0.6.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
boto3==1.34.34

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0